requirements.txt
django==5.1.4 
djangorestframework==3.15.2
geopy==2.4.1 
timezonefinder==6.5.7
requests
aiohttp
python-dotenv
pytz
swisseph
kerykeion
pyswisseph


 
